+++ src/form/from-group-footer/form-group-footer.component.ts
@@ -3,14 +3,14 @@ import {
     HostBinding,
     ViewEncapsulation,
     OnInit,
-    ContentChild,
     TemplateRef,
     inject,
     input,
-    ChangeDetectionStrategy
+    ChangeDetectionStrategy,
+    contentChild
 } from '@angular/core';
 import { ThyFormDirective } from '../form.directive';
-import { ThyFormGroupFooterAlign, THY_FORM_CONFIG, ThyFormConfig } from '../form.class';
+import { ThyFormGroupFooterAlign, THY_FORM_CONFIG } from '../form.class';
 import { NgClass, NgTemplateOutlet } from '@angular/common';
 
 /**
@@ -33,7 +33,7 @@ export class ThyFormGroupFooter implements OnInit {
 
     @HostBinding('class.row') isHorizontal = true;
 
-    @ContentChild('description') description?: TemplateRef<any>;
+    readonly description = contentChild<TemplateRef<any>>('description');
 
     /**
      * 对齐方式