+++ src/public-api.ts
@@ -1,1 +1,3 @@
 export * from './version';
+export * from './core/global-config';
+export * from './core/provide';